@@ -11,17 +11,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.0.0",
-    "@angular/cdk": "^12.2.13",
-    "@angular/common": "~13.0.2",
-    "@angular/compiler": "~13.0.0",
-    "@angular/core": "~13.0.0",
-    "@angular/fire": "^7.0.4",
-    "@angular/forms": "~13.0.0",
-    "@angular/material": "^12.2.13",
-    "@angular/platform-browser": "~13.0.0",
-    "@angular/platform-browser-dynamic": "~13.0.0",
-    "@angular/router": "~13.0.0",
+    "@angular/animations": "^14.0.1",
+    "@angular/cdk": "^13.3.9",
+    "@angular/common": "^14.0.1",
+    "@angular/compiler": "^14.0.1",
+    "@angular/core": "^14.0.1",
+    "@angular/fire": "^7.4.1",
+    "@angular/forms": "^14.0.1",
+    "@angular/material": "^13.3.9",
+    "@angular/platform-browser": "^14.0.1",
+    "@angular/platform-browser-dynamic": "^14.0.1",
+    "@angular/router": "^14.0.1",
     "angular-google-charts": "^2.2.2",
     "bootstrap": "^4.6.1",
     "cors": "^2.8.5",
@@ -43,9 +43,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.0.3",
-    "@angular/cli": "~13.0.3",
-    "@angular/compiler-cli": "~13.0.0",
+    "@angular-devkit/build-angular": "^14.0.1",
+    "@angular/cli": "^14.0.1",
+    "@angular/compiler-cli": "^14.0.1",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
     "@types/xml2js": "^0.4.8",
@@ -66,6 +66,6 @@
     "open": "^7.0.3",
     "protractor": "^7.0.0",
     "ts-node": "~8.3.0",
-    "typescript": "~4.4.3"
+    "typescript": "~4.7.3"
   }
 }
