@@ -12,7 +12,7 @@
 * See the MIT License for more details. 
 ***********************************************************/
 import { Injectable } from '@angular/core';
-import { FormControl, Validators } from '@angular/forms';
+import { UntypedFormControl, Validators } from '@angular/forms';
 import { ErrorHandlerService } from '../error-handler/error-handler.service';
 
 @Injectable({
@@ -163,7 +163,7 @@ export class ValidationService {
     }
 
     async checkTitle(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -174,7 +174,7 @@ export class ValidationService {
     }
 
     async checkStatus(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -185,7 +185,7 @@ export class ValidationService {
     }
 
     async checkPriority(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -196,7 +196,7 @@ export class ValidationService {
     }
 
     async checkEstimatedTime(value: number) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -207,7 +207,7 @@ export class ValidationService {
     }
 
     async checkDifficulty(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -218,7 +218,7 @@ export class ValidationService {
     }
 
     async checkDescription(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -229,7 +229,7 @@ export class ValidationService {
     }
 
     async checkCreator(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -240,7 +240,7 @@ export class ValidationService {
     }
 
     async checkProject(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -251,7 +251,7 @@ export class ValidationService {
     }
 
     async checkAssignee(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         const regularExpression = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
         const result = regularExpression.test(String(value).toLowerCase());
         if (control.errors === null && result == true)
@@ -264,7 +264,7 @@ export class ValidationService {
     }
 
     async checkReporter(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         const regularExpression = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
         const result = regularExpression.test(String(value).toLowerCase());
         if (control.errors === null && result == true)
@@ -277,7 +277,7 @@ export class ValidationService {
     }
 
     async checkCreationDate(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -288,7 +288,7 @@ export class ValidationService {
     }
 
     async checkSprintNumber(value: number) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -299,7 +299,7 @@ export class ValidationService {
     }
 
     async checkStoryPoint(value: number) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -310,7 +310,7 @@ export class ValidationService {
     }
 
     async checkStartDate(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         const regularExpression = /^\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$/;
         const result = regularExpression.test(String(value));
         if (control.errors === null && result == true)
@@ -325,7 +325,7 @@ export class ValidationService {
     }
 
     async checkEndDate(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         const regularExpression = /^\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$/;
         const result = regularExpression.test(String(value));
         if (control.errors === null && result == true)
@@ -338,7 +338,7 @@ export class ValidationService {
     }
 
     async checkLogHours(value: number) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -349,7 +349,7 @@ export class ValidationService {
     }
 
     async checkWorkCompleted(value: number) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -360,7 +360,7 @@ export class ValidationService {
     }
 
     async checkComment(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -395,7 +395,7 @@ export class ValidationService {
         return (true)
     }
     async checkUserEmail(value: String) {
-        const control = new FormControl(value, [Validators.required, Validators.email]);
+        const control = new UntypedFormControl(value, [Validators.required, Validators.email]);
         if (control.errors === null)
             return (true);
         else {
@@ -405,7 +405,7 @@ export class ValidationService {
         }
     }
     async checkUserName(value: String) {
-        const control = new FormControl(value, [Validators.required, Validators.minLength(3)]);
+        const control = new UntypedFormControl(value, [Validators.required, Validators.minLength(3)]);
         if (control.errors === null)
             return (true);
         else {
@@ -415,7 +415,7 @@ export class ValidationService {
         }
     }
     async checkUserPassword(value: string) {
-        const control = new FormControl(value, [Validators.required, Validators.minLength(6)]);
+        const control = new UntypedFormControl(value, [Validators.required, Validators.minLength(6)]);
         if (control.errors === null)
             return (true);
         else {
@@ -425,7 +425,7 @@ export class ValidationService {
         }
     }
     async checkOrgName(value: string) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -435,7 +435,7 @@ export class ValidationService {
         }
     }
     async checkOrgDomain(value: string) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -445,7 +445,7 @@ export class ValidationService {
         }
     }
     async checkOrgEmail(value: string) {
-        const control = new FormControl(value, [Validators.required, Validators.email]);
+        const control = new UntypedFormControl(value, [Validators.required, Validators.email]);
         if (control.errors === null)
             return (true);
         else {
@@ -455,7 +455,7 @@ export class ValidationService {
         }
     }
     async checkOrgLogo(value: string) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -465,7 +465,7 @@ export class ValidationService {
         }
     }
     async checkTeamName(value: string) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -475,7 +475,7 @@ export class ValidationService {
         }
     }
     async checkTeamId(value: string) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -485,7 +485,7 @@ export class ValidationService {
         }
     }
     async checkTeamDescription(value: string) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -495,7 +495,7 @@ export class ValidationService {
         }
     }
     async checkTeamManagerEmail(value: string) {
-        const control = new FormControl(value, [Validators.required, Validators.email]);
+        const control = new UntypedFormControl(value, [Validators.required, Validators.email]);
         if (control.errors === null)
             return (true);
         else {
@@ -507,7 +507,7 @@ export class ValidationService {
     }
 
     async checkinstituteName(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -518,7 +518,7 @@ export class ValidationService {
     }
 
     async checkdegree(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -529,7 +529,7 @@ export class ValidationService {
     }
      
     async checkprojectName(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -540,7 +540,7 @@ export class ValidationService {
     }
 
     async checkdescription(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -551,7 +551,7 @@ export class ValidationService {
     }
 
     async checkskill(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -562,7 +562,7 @@ export class ValidationService {
     }
 
     async checkposition(value: String) {
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         if (control.errors === null)
             return (true);
         else {
@@ -587,7 +587,7 @@ export class ValidationService {
             return (false);
         }
         value.map((member, index) => {
-            const control = new FormControl(member, [Validators.required, Validators.email]);
+            const control = new UntypedFormControl(member, [Validators.required, Validators.email]);
             if (control.errors === null)
                 hasNoError = true
             else {
@@ -599,7 +599,7 @@ export class ValidationService {
     }
 
     async checkURL(value: string[]){
-        const control = new FormControl(value, Validators.required);
+        const control = new UntypedFormControl(value, Validators.required);
         const regularExpression = /^((https?|ftp|smtp):\/\/)?(www.)?[a-z0-9]+\.[a-z]+(\/[a-zA-Z0-9#]+\/?)*$/;
         const result = regularExpression.test(String(value).toLowerCase());
         if (control.errors === null && result == true)
